Move CCS to training folder
